basic server threading working

+++ frontend/cuckoo/SimpleCuckoo.cpp
@@ -126,7 +126,7 @@ namespace osuCrypto
 
         mHashesView = MatrixView<u64>(mHashes.begin(), mHashes.end(), mParams.mNumHashes);
 
-        u64 binCount = mParams.mBinScaler * n;
+        u64 binCount = u64(mParams.mBinScaler * n);
 
         mBins.resize(binCount);
         mStash.resize(mParams.mStashSize);
