+++ libPSI/Tools/CuckooHasher.cpp
@@ -150,7 +150,7 @@ namespace osuCrypto
 
         mHashesView = MatrixView<u64>(mHashes.begin(), mHashes.end(), mParams.mNumHashes);
 
-        u64 binCount = mParams.mBinScaler * mParams.mN;
+        u64 binCount = u64(mParams.mBinScaler * mParams.mN);
 
         mBins.resize(binCount);
         mStash.resize(mParams.mStashSize);
