+++ libPSI_Tests/OTOracleSender.cpp
@@ -30,7 +30,7 @@ void OTOracleSender::send(
 
     u64 doneIdx = 0;
 
-    for (doneIdx = 0; doneIdx < messages.size(); ++doneIdx)
+    for (doneIdx = 0; doneIdx < u64(messages.size()); ++doneIdx)
     {
         messages[doneIdx][0] = mPrng.get<block>();
         messages[doneIdx][1] = mPrng.get<block>();
